chore(server): modulate schema definitions
