+++ server/package.json
@@ -14,7 +14,9 @@
   "type": "module",
   "dependencies": {
     "@apollo/server": "^4.9.0",
-    "graphql": "^16.7.1"
+    "@graphql-tools/schema": "^10.0.0",
+    "graphql": "^16.7.1",
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
     "@types/node": "^20.4.5",
