+++ server/src/definitions/index.ts
@@ -1,0 +1,8 @@
+import { makeExecutableSchema } from "@graphql-tools/schema";
+import { mergeResolvers, mergeTypeDefs } from "@graphql-tools/merge";
+import { typeDefs as Books, resolvers as ResolverBooks } from "./books.js";
+
+export const schema = makeExecutableSchema({
+  typeDefs: mergeTypeDefs([Books]),
+  resolvers: mergeResolvers([ResolverBooks]),
+});
