+++ server/tsconfig.json
@@ -13,6 +13,8 @@
         "esModuleInterop": true,
         "types": [
             "node"
-        ]
+        ],
+        "allowImportingTsExtensions": true,
+        "noEmit": true
     }
 }
